swn: moved config to pkg/config
