+++ pkg/swn/p2p/peer_test.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"go.neonyx.io/go-swn/pkg/config"
 	"go.neonyx.io/go-swn/pkg/logger"
-	"go.neonyx.io/go-swn/pkg/swn/config"
 	"go.neonyx.io/go-swn/pkg/swn/p2p"
 )
